[REFACTOR] Friend URL 구조 변경 및 일부 API User 컨트롤러로 이동
